chore: 공통 컴포넌트 함수 정의 위치 변경
