+++ src/common/Groups.tsx
@@ -1,102 +1,21 @@
-import React, { useRef, useState } from 'react';
+import React from 'react';
 import styled, { css } from 'styled-components';
-import {
-  IcArrowLeftSmallGray,
-  IcArrowRightSmallGray,
-  IcStatusBlack,
-  IcStatusWhite,
-  IcSuccess,
-  IcSuccessGray,
-} from '../assets';
-import { SORTING, STATUS } from '../constants/Follower/currentConst';
+import { IcArrowLeftSmallGray, IcArrowRightSmallGray } from '../assets';
 import { PersonalGroupProps } from '../types/MyGroup/myGroupType';
 import RecommendCard from './RecommendCard';
 
-const Groups = ({ group }: PersonalGroupProps) => {
-  // 일단 더미 데이터로 넣어둠, 서버 연결 시 0으로 초기화하고 서버에서 가져온 전체 페이지 데이터로 업데이트 예정
-  const totalPageRef = useRef(3);
-  const pages = Array.from(
-    { length: totalPageRef.current },
-    (_, idx) => idx + 1
-  );
-
-  const [filter, setFilter] = useState({
-    clickedStatus: '모집 중',
-    sorting: '최신순',
-  });
-  const [clickedPage, setClickedPage] = useState(1);
-
-  const { clickedStatus, sorting } = filter;
-
-  const handleClickSorting = (
-    e:
-      | React.MouseEvent<HTMLParagraphElement, MouseEvent>
-      | React.MouseEvent<HTMLButtonElement, MouseEvent>,
-    isSorting: boolean
-  ) => {
-    const { innerText } = e.currentTarget;
-
-    isSorting
-      ? setFilter({
-          ...filter,
-          sorting: innerText,
-        })
-      : setFilter({
-          ...filter,
-          clickedStatus: innerText,
-        });
-    // 최신순/ 가나다순에 따라 서버 통신 들어갈 예정
-  };
-
-  const handleClickPrevBtn = () => {
-    setClickedPage((prev) => prev - 1);
-  };
-
-  const handleClickPage = (page: number) => {
-    setClickedPage(page);
-  };
-
-  const handleClickNextBtn = () => {
-    setClickedPage((prev) => prev + 1);
-  };
+const Groups = ({
+  group,
+  totalPage,
+  clickedPage,
+  handleClickPages,
+}: PersonalGroupProps) => {
+  const pages = Array.from({ length: totalPage }, (_, idx) => idx + 1);
+  const { handleClickPrevBtn, handleClickPage, handleClickNextBtn } =
+    handleClickPages;
 
   return (
     <React.Fragment>
-      <TopContainer>
-        <TotalStatus>
-          {STATUS.map((status, idx) => {
-            return (
-              <StatusContainer key={status}>
-                {clickedStatus === status ? <IcSuccess /> : <IcSuccessGray />}
-
-                <Status
-                  type="button"
-                  $idx={idx}
-                  onClick={(e) => handleClickSorting(e, false)}
-                >
-                  {idx === 0 ? <IcStatusBlack /> : <IcStatusWhite />}
-                  <Text $idx={idx}>{status}</Text>
-                </Status>
-              </StatusContainer>
-            );
-          })}
-        </TotalStatus>
-
-        <SortContainer>
-          {SORTING.map((standard) => {
-            return (
-              <Sorting
-                key={standard}
-                onClick={(e) => standard !== '|' && handleClickSorting(e, true)}
-                $isClicked={sorting === standard}
-              >
-                {standard}
-              </Sorting>
-            );
-          })}
-        </SortContainer>
-      </TopContainer>
-
       <RecommendCard group={group} isLongPage={true} />
 
       <PageNationBar>
@@ -115,9 +34,7 @@ const Groups = ({ group }: PersonalGroupProps) => {
           );
         })}
         <IcArrowRightSmallGray
-          onClick={() =>
-            clickedPage !== totalPageRef.current && handleClickNextBtn()
-          }
+          onClick={() => clickedPage !== totalPage && handleClickNextBtn()}
         />
       </PageNationBar>
     </React.Fragment>
@@ -125,68 +42,6 @@ const Groups = ({ group }: PersonalGroupProps) => {
 };
 
 export default Groups;
-
-const TopContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  width: 100%;
-  margin: 1.4rem 0.1rem 0 0.6rem;
-`;
-
-const TotalStatus = styled.div`
-  display: flex;
-  gap: 1.4rem;
-  justify-content: center;
-  align-items: center;
-`;
-
-const StatusContainer = styled.div`
-  display: flex;
-  gap: 0.6rem;
-  justify-content: center;
-  align-items: center;
-`;
-
-const Status = styled.button<{ $idx: number }>`
-  display: flex;
-  gap: 0.8rem;
-  justify-content: center;
-  align-items: center;
-
-  padding: 0.7rem 1.4rem 0.7rem 1.2rem;
-
-  border-radius: 0.6rem;
-  background-color: ${({ $idx, theme }) => {
-    switch ($idx) {
-      case 0:
-        return `${theme.colors.codrive_green}`;
-      case 1:
-        return `${theme.colors.codrive_purple}`;
-      case 2:
-        return `${theme.colors.gray600}`;
-    }
-  }};
-`;
-
-const Text = styled.p<{ $idx: number }>`
-  color: ${({ $idx, theme }) =>
-    $idx === 0 ? theme.colors.gray900 : theme.colors.white};
-  ${({ theme }) => theme.fonts.title_bold_14};
-`;
-
-const SortContainer = styled.div`
-  display: flex;
-  gap: 0.8rem;
-  align-items: center;
-`;
-
-const Sorting = styled.p<{ $isClicked: boolean }>`
-  color: ${({ $isClicked, theme }) =>
-    $isClicked ? theme.colors.white : theme.colors.gray500};
-  ${({ theme }) => theme.fonts.body_medium_14};
-`;
 
 const PageNationBar = styled.div`
   display: flex;
